pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "wtfd"
version = "0.1.0"
description = "Wind Turbine Fault Detection using SCADA telemetry data"
authors = [
    {name = "Curtis Neiderer"}
]
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
    "numpy",
    "pandas",
    "scikit-learn",
    "matplotlib",
    "seaborn",
    "pyyaml"
]

[tool.setuptools.packages.find]
where = ["src"]